stap 3 is af
