+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/nynke/Documents/Opleiding Full Stack Developer/Frontend/frontend-react-pokemon/src/index.js":"1","/Users/nynke/Documents/Opleiding Full Stack Developer/Frontend/frontend-react-pokemon/src/reportWebVitals.js":"2","/Users/nynke/Documents/Opleiding Full Stack Developer/Frontend/frontend-react-pokemon/src/App.js":"3"},{"size":478,"mtime":1653400503720,"results":"4","hashOfConfig":"5"},{"size":362,"mtime":1653400503720,"results":"6","hashOfConfig":"5"},{"size":1466,"mtime":1653407746211,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"10"},"1x0ioe6",{"filePath":"11","messages":"12","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"10"},{"filePath":"13","messages":"14","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/Users/nynke/Documents/Opleiding Full Stack Developer/Frontend/frontend-react-pokemon/src/index.js",[],["15","16"],"/Users/nynke/Documents/Opleiding Full Stack Developer/Frontend/frontend-react-pokemon/src/reportWebVitals.js",[],"/Users/nynke/Documents/Opleiding Full Stack Developer/Frontend/frontend-react-pokemon/src/App.js",["17","18"],{"ruleId":"19","replacedBy":"20"},{"ruleId":"21","replacedBy":"22"},{"ruleId":"23","severity":1,"message":"24","line":5,"column":14,"nodeType":"25","messageId":"26","endLine":5,"endColumn":16},{"ruleId":"27","severity":1,"message":"28","line":33,"column":29,"nodeType":"29","endLine":33,"endColumn":35},"no-native-reassign",["30"],"no-negated-in-lhs",["31"],"no-empty-pattern","Unexpected empty object pattern.","ObjectPattern","unexpected","jsx-a11y/alt-text","img elements must have an alt prop, either with meaningful text, or an empty string for decorative images.","JSXOpeningElement","no-global-assign","no-unsafe-negation"]
+[{"/Users/nynke/Documents/Opleiding Full Stack Developer/Frontend/frontend-react-pokemon/src/index.js":"1","/Users/nynke/Documents/Opleiding Full Stack Developer/Frontend/frontend-react-pokemon/src/reportWebVitals.js":"2","/Users/nynke/Documents/Opleiding Full Stack Developer/Frontend/frontend-react-pokemon/src/App.js":"3","/Users/nynke/Documents/Opleiding Full Stack Developer/Frontend/frontend-react-pokemon/src/components/Pokemon.js":"4"},{"size":478,"mtime":1653400503720,"results":"5","hashOfConfig":"6"},{"size":362,"mtime":1653400503720,"results":"7","hashOfConfig":"6"},{"size":292,"mtime":1653474147881,"results":"8","hashOfConfig":"6"},{"size":1317,"mtime":1653474503584,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"1x0ioe6",{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"15","messages":"16","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"17","messages":"18","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/Users/nynke/Documents/Opleiding Full Stack Developer/Frontend/frontend-react-pokemon/src/index.js",[],["19","20"],"/Users/nynke/Documents/Opleiding Full Stack Developer/Frontend/frontend-react-pokemon/src/reportWebVitals.js",[],"/Users/nynke/Documents/Opleiding Full Stack Developer/Frontend/frontend-react-pokemon/src/App.js",["21","22"],"/Users/nynke/Documents/Opleiding Full Stack Developer/Frontend/frontend-react-pokemon/src/components/Pokemon.js",["23"],{"ruleId":"24","replacedBy":"25"},{"ruleId":"26","replacedBy":"27"},{"ruleId":"28","severity":1,"message":"29","line":1,"column":16,"nodeType":"30","messageId":"31","endLine":1,"endColumn":24},{"ruleId":"28","severity":1,"message":"32","line":3,"column":8,"nodeType":"30","messageId":"31","endLine":3,"endColumn":13},{"ruleId":"33","severity":1,"message":"34","line":31,"column":21,"nodeType":"35","endLine":31,"endColumn":67},"no-native-reassign",["36"],"no-negated-in-lhs",["37"],"no-unused-vars","'useState' is defined but never used.","Identifier","unusedVar","'axios' is defined but never used.","jsx-a11y/alt-text","img elements must have an alt prop, either with meaningful text, or an empty string for decorative images.","JSXOpeningElement","no-global-assign","no-unsafe-negation"]
